Refactor blog and user functionalities for better user experience
This commit prioritizes user experience by adding a form for blog posts and enhancing security in the user model. The login and registration processes are improved and blog form is now part of the blog view. The form handlebars file is moved to a partials directory and modifications are performed in the script.js, home-routes.js, user.js files for optimized functionality. Styles have also been updated in blog-post.scss for form accessibility.